+++ CHANGELOG.md
@@ -9,8 +9,9 @@ All notable changes to the `dom_query` crate will be documented in this file.
 - `Selection::remove` method now does not require `&mut`.
 
 ## Added
+- Added `Node::append_html` and `Document::set_html` methods for creating children nodes of a single selected node.
 - Added `Tree<NodeData>::new_element`, an easy way to create an empty element with a given name.
-- Added `NodeRef::last_child`;
+- Added `NodeRef::last_child`.
 - Added `Node::has_attr` method, which returns true if an attribute exists on the node element. 
 `Selection::has_attr` does the same thing for the first element inside selection.
 - Added `Selection::remove_all_attrs` method, a way to remove all attributes from all elements in the selection.